Fix missing base_distance in snow cache update
The updateCache() function was not including the base_distance column
in the INSERT statement, causing a NOT NULL constraint violation:

  ERROR: null value in column "base_distance" of relation
  "snow_totals_cache" violates not-null constraint (SQLSTATE 23502)

Added base_distance to both the INSERT column list and the DO UPDATE
clause, passing c.baseDistance as the 6th parameter.

This fixes the 30-second snow totals cache refresh cycle.